Trying to fix a weird bug due to malloc

+++ src/analyzer/analyzer.h
@@ -10,7 +10,7 @@
  */
 typedef struct FileInfo{
     char *name;
-    int *fileTable;
+    unsigned long long *fileTable;
     char *path;
     int isDirectory;
 } *FileInfo;
